Correct minor bugs (unsigned buffer for input file and size of integer for vector elements)

+++ insertion_sort/insertion_sort.cpp
@@ -2,17 +2,18 @@
 
 #include "insertion_sort.h"
 
-void Sort::insertion (std::vector<int>& v) {
+void Sort::insertion (std::vector<long>& v) {
    unsigned int vsize;
-   int aux;
+   long aux;
    int i;
 
    vsize = v.size();
 
-   for (unsigned int sorted = 0; sorted < vsize-1; sorted++) {
+   for (unsigned sorted = 0; sorted < vsize-1; sorted++) {
       aux = v[sorted+1];
       for (i = sorted; (i >= 0) && (aux < v[i]); i--)
          v[i+1] = v[i];
       v[i+1] = aux;
    }
 }
+
