+++ merge_sort/merge_sort.h
@@ -7,8 +7,8 @@
 # include <bits/stdc++.h>
 
 namespace Sort {
-   void mergeSort (std::vector<int>& v, unsigned l, unsigned r);
-   void merge (std::vector<int>& v, unsigned l, unsigned m, unsigned r);
+   void mergeSort (std::vector<long>& v, unsigned l, unsigned r);
+   void merge (std::vector<long>& v, unsigned l, unsigned m, unsigned r);
 }
 
 #endif
